Refactor : common 패키지 생성하여, product와 user가 공통으로 관련있는 클래스를 common 패키지로 옮김
