+++ app/src/main/java/com/codesoom/assignment/common/exception/ProductNotFoundException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment;
+package com.codesoom.assignment.common.exception;
 
 public class ProductNotFoundException extends RuntimeException {
     public ProductNotFoundException(Long id) {
