+++ app/src/test/java/com/codesoom/assignment/product/application/ProductServiceTest.java
@@ -1,10 +1,9 @@
 package com.codesoom.assignment.product.application;
 
-import com.codesoom.assignment.ProductNotFoundException;
+import com.codesoom.assignment.common.exception.ProductNotFoundException;
 import com.codesoom.assignment.product.domain.Product;
 import com.codesoom.assignment.product.domain.ProductRepository;
 import com.codesoom.assignment.product.dto.ProductData;
-import com.github.dozermapper.core.DozerBeanMapper;
 import com.github.dozermapper.core.DozerBeanMapperBuilder;
 import com.github.dozermapper.core.Mapper;
 import org.junit.jupiter.api.BeforeEach;
